fix: cambios en los estilos de la pagina de login
