+++ FRONTEND/README.md
@@ -1,8 +1,21 @@
-# React + Vite
+# FUNCIONES DE LA APLICACIÓN
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+# Páginas
 
-Currently, two official plugins are available:
+LoginPage: Formulario para poder ingresar o iniciar sesión si se tiene un usuario registrado, los usuarios solo pueden ser agregados por un usuario con rol de administrador
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+# ESTILOS TAILWIND
+
+# Componentes
+
+LoginInfo
+
+h-full w-full: Asegura que la imagen ocupe todo el alto y ancho de su contenedor.
+
+rounded-lg: Aplica bordes redondeados al contenedor y a la imagen.
+
+overflow-hidden: Asegura que cualquier parte de la imagen que sobresalga quede oculta dentro de los bordes redondeados.
+
+object-cover: Hace que la imagen mantenga su proporción y ocupe todo el espacio disponible en el contenedor sin distorsionarse.
+
+shadow-lg: Añade una sombra al contenedor de la imagen para darle un efecto visual.
