+++ FRONTEND/tailwind.config.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   content: [
     "./index.html",
     "./src/**/*.{js,jsx,ts,tsx}",
@@ -7,4 +7,4 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-}
+};
